refactor(collie-core): split the CollieDB god object into storage domains
collie_core/db.py had grown to 4,072 lines: one class holding 164 methods
spanning conversations, life tools, reminders, shopping, connectors,
providers, checklists and approvals. Finding one domain meant reading past
everything else, and every feature branch touched the same file.

Each storage domain now lives in one module under collie_core/db_domains/
and is mixed into CollieDB, which stays the single public interface for
storage, so no call site changes: db.<method> keeps working for all 77
importers of CollieDB.

db_primitives.py (collie_home, utc_now, new_id) moved to a leaf module
because db.py imports the domains, so a domain importing db.py back for
new_id would be an import cycle. db.py imports and re-exports those names,
so from collie_core.db import collie_home keeps working. _PROVIDER_SETTING_KEYS
moved to ProvidersDomain, which was its only reader.

Behaviour is unchanged. Every one of the 164 members was compared with
inspect.getsource before and after the move: 0 missing, 0 changed. The full
backend suite passes (3,768 passed, 2 skipped), the ruff lint and format
gates pass, and the coverage gate holds.

tests/collie/test_db_domain_split.py keeps the split honest: it pins which
domain owns which method, rejects a name owned by two domains (a silent MRO
shadow), and rejects a domain that reaches for another domain or for
anything outside the host plumbing.

db.py is now 2,091 lines. The run, plan, checklist-revision and telemetry
methods stay in db.py on purpose: they share _write_immediate transaction
semantics and the run-task revision helpers, so they are one transactional
cluster, not a set of domains.

docs/engineering/architecture/storage-domains.md records the seam and how to
add a domain; PROJECT_MAP.md and the repository snapshot are updated.
