+++ collie-core/collie_core/db.py
@@ -8,6 +8,9 @@ Design:
 - One writer connection guarded by a lock; WAL mode so readers never block.
 - ``schema_version`` table + ordered migration list for upgrades.
 - All timestamps are ISO-8601 UTC strings.
+- Connection plumbing, schema and migrations live in this module. One storage
+  domain per module lives in :mod:`collie_core.db_domains`; those modules are
+  mixed into :class:`CollieDB`, so ``db.<method>`` stays the public interface.
 """
 
 from __future__ import annotations
@@ -17,28 +20,26 @@ import json
 import os
 import sqlite3
 import threading
-import uuid
 from collections.abc import Iterator
-from contextlib import contextmanager, suppress
-from datetime import UTC, date, datetime, timedelta
+from contextlib import contextmanager
+from datetime import date
 from pathlib import Path
 from typing import Any
 
-__all__ = ["CollieDB", "collie_home", "utc_now"]
+from collie_core.db_domains import (
+    ApprovalsDomain,
+    ArtifactsDomain,
+    AutomationsDomain,
+    ChecklistsDomain,
+    ConnectorsDomain,
+    ConversationsDomain,
+    LifeDomain,
+    ProvidersDomain,
+    SettingsDomain,
+)
+from collie_core.db_primitives import collie_home, new_id, utc_now
 
-
-def collie_home() -> Path:
-    """Return the Collie data directory (``~/.collie`` or ``$COLLIE_HOME``)."""
-    root = os.environ.get("COLLIE_HOME")
-    return Path(root).expanduser() if root else Path.home() / ".collie"
-
-
-def utc_now() -> str:
-    return datetime.now(UTC).isoformat(timespec="seconds")
-
-
-def new_id() -> str:
-    return uuid.uuid4().hex
+__all__ = ["CollieDB", "collie_home", "new_id", "utc_now"]
 
 
 _SCHEMA_V1 = """
@@ -678,8 +679,23 @@ _MIGRATIONS: list[str] = [
 ]
 
 
-class CollieDB:
-    """Thread-safe SQLite wrapper for all Collie structured storage."""
+class CollieDB(
+    SettingsDomain,
+    ConversationsDomain,
+    LifeDomain,
+    AutomationsDomain,
+    ChecklistsDomain,
+    ApprovalsDomain,
+    ConnectorsDomain,
+    ProvidersDomain,
+    ArtifactsDomain,
+):
+    """Thread-safe SQLite wrapper for all Collie structured storage.
+
+    The connection, schema, migrations and shared row helpers live here; each
+    storage domain is mixed in from :mod:`collie_core.db_domains` and owns its
+    own tables. Callers keep using ``db.<method>`` for everything.
+    """
 
     def __init__(self, db_path: Path | None = None):
         self.path = db_path or (collie_home() / "collie.db")
@@ -779,1148 +795,12 @@ class CollieDB:
             row = cur.fetchone()
             return dict(row) if row else None
 
-    # -- settings --------------------------------------------------------------
-
-    def get_setting(self, key: str, default: Any = None) -> Any:
-        row = self._row("SELECT value FROM settings WHERE key = ?", (key,))
-        if row is None:
-            return default
-        try:
-            return json.loads(row["value"])
-        except (TypeError, ValueError):
-            return row["value"]
-
-    def set_setting(self, key: str, value: Any) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO settings (key, value) VALUES (?, ?) "
-                "ON CONFLICT(key) DO UPDATE SET value = excluded.value",
-                (key, json.dumps(value)),
-            )
-
-    def set_active_model(self, model: str) -> None:
-        """Persist the active model atomically: the global setting AND the
-        default provider row stay in agreement (one canonical source)."""
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO settings (key, value) VALUES (?, ?) "
-                "ON CONFLICT(key) DO UPDATE SET value = excluded.value",
-                ("provider.model", json.dumps(model)),
-            )
-            conn.execute(
-                "UPDATE providers SET model = ? WHERE is_default = 1",
-                (model,),
-            )
-
-    def all_settings(self) -> dict[str, Any]:
-        out: dict[str, Any] = {}
-        for row in self._rows("SELECT key, value FROM settings"):
-            try:
-                out[row["key"]] = json.loads(row["value"])
-            except (TypeError, ValueError):
-                out[row["key"]] = row["value"]
-        return out
-
-    def delete_setting(self, key: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM settings WHERE key = ?", (key,))
-
-    # -- conversations -----------------------------------------------------------
-
-    def create_conversation(
-        self,
-        title: str = "New chat",
-        conv_id: str | None = None,
-        project_path: str | None = None,
-    ) -> dict[str, Any]:
-        cid = conv_id or new_id()
-        now = utc_now()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO conversations "
-                "(id, title, created_at, updated_at, project_path) VALUES (?, ?, ?, ?, ?)",
-                (cid, title, now, now, project_path),
-            )
-        return {
-            "id": cid,
-            "title": title,
-            "created_at": now,
-            "updated_at": now,
-            "archived": 0,
-            "project_path": project_path,
-        }
-
-    def get_conversation(self, conv_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM conversations WHERE id = ?", (conv_id,))
-
-    def set_conversation_mode(self, conv_id: str, mode: str) -> None:
-        if mode not in {"plan", "execute"}:
-            raise ValueError("execution mode must be 'plan' or 'execute'")
-        with self._write() as conn:
-            conn.execute(
-                "UPDATE conversations SET execution_mode = ?, updated_at = ? WHERE id = ?",
-                (mode, utc_now(), conv_id),
-            )
-
-    def set_conversation_project(self, conv_id: str, project_path: str | None) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "UPDATE conversations SET project_path = ?, updated_at = ? WHERE id = ?",
-                (project_path, utc_now(), conv_id),
-            )
-
-    def list_conversations(self, include_archived: bool = False) -> list[dict[str, Any]]:
-        if include_archived:
-            return self._rows("SELECT * FROM conversations ORDER BY updated_at DESC")
-        return self._rows("SELECT * FROM conversations WHERE archived = 0 ORDER BY updated_at DESC")
-
-    def rename_conversation(self, conv_id: str, title: str) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "UPDATE conversations SET title = ?, updated_at = ? WHERE id = ?",
-                (title, utc_now(), conv_id),
-            )
-
-    def delete_conversation(self, conv_id: str) -> None:
-        with self._write() as conn:
-            # Children first (messages has an FK to conversations; run_steps
-            # cascades from runs, approval_requests is orphaned otherwise).
-            conn.execute(
-                "DELETE FROM task_checklist_steps WHERE checklist_id IN "
-                "(SELECT id FROM task_checklists WHERE conversation_id = ?)",
-                (conv_id,),
-            )
-            conn.execute("DELETE FROM task_checklists WHERE conversation_id = ?", (conv_id,))
-            conn.execute(
-                "DELETE FROM conversation_review_gates WHERE conversation_id = ?", (conv_id,)
-            )
-            conn.execute("DELETE FROM plan_change_requests WHERE conversation_id = ?", (conv_id,))
-            conn.execute(
-                "DELETE FROM run_steps WHERE run_id IN "
-                "(SELECT id FROM runs WHERE conversation_id = ?)",
-                (conv_id,),
-            )
-            conn.execute("DELETE FROM runs WHERE conversation_id = ?", (conv_id,))
-            conn.execute("DELETE FROM plans WHERE conversation_id = ?", (conv_id,))
-            conn.execute("DELETE FROM approval_requests WHERE conversation_id = ?", (conv_id,))
-            conn.execute("DELETE FROM messages WHERE conversation_id = ?", (conv_id,))
-            conn.execute("DELETE FROM conversations WHERE id = ?", (conv_id,))
-
-    # -- messages -----------------------------------------------------------------
-
-    def add_message(
-        self,
-        conversation_id: str,
-        role: str,
-        content: str,
-        *,
-        tool_calls: Any = None,
-        tool_results: Any = None,
-        card_type: str | None = None,
-        card_data: Any = None,
-        task_state: dict[str, Any] | None = None,
-        attachments: Any = None,
-        token_count: int | None = None,
-        msg_id: str | None = None,
-    ) -> dict[str, Any]:
-        mid = msg_id or new_id()
-        now = utc_now()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO messages (id, conversation_id, role, content, tool_calls, "
-                "tool_results, card_type, card_data, task_state, attachments, token_count, "
-                "created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                (
-                    mid,
-                    conversation_id,
-                    role,
-                    content,
-                    json.dumps(tool_calls) if tool_calls is not None else None,
-                    json.dumps(tool_results) if tool_results is not None else None,
-                    card_type,
-                    json.dumps(card_data) if card_data is not None else None,
-                    json.dumps(task_state) if task_state is not None else None,
-                    json.dumps(attachments) if attachments is not None else None,
-                    token_count,
-                    now,
-                ),
-            )
-            conn.execute(
-                "UPDATE conversations SET updated_at = ? WHERE id = ?",
-                (now, conversation_id),
-            )
-        return {
-            "id": mid,
-            "conversation_id": conversation_id,
-            "role": role,
-            "content": content,
-            "card_type": card_type,
-            "card_data": card_data,
-            "task_state": task_state,
-            "attachments": attachments,
-            "created_at": now,
-        }
-
-    def get_messages(self, conversation_id: str, limit: int | None = None) -> list[dict[str, Any]]:
-        sql = "SELECT * FROM messages WHERE conversation_id = ? ORDER BY created_at, rowid"
-        params: tuple = (conversation_id,)
-        if limit is not None:
-            if limit <= 0:
-                return []
-            count = self._row(
-                "SELECT COUNT(*) AS n FROM messages WHERE conversation_id = ?",
-                (conversation_id,),
-            )
-            offset = max(0, int(count["n"]) - limit)
-            sql += " LIMIT ? OFFSET ?"
-            params = (conversation_id, limit, offset)
-        rows = self._rows(sql, params)
-        for r in rows:
-            for field in ("tool_calls", "tool_results", "card_data", "task_state", "attachments"):
-                if r.get(field):
-                    with suppress(TypeError, ValueError):
-                        r[field] = json.loads(r[field])
-        return rows
-
-    def all_messages_with_attachments(self) -> list[dict[str, Any]]:
-        """Every message that references stored media, with attachments parsed."""
-        rows = self._rows(
-            "SELECT conversation_id, attachments FROM messages WHERE attachments IS NOT NULL"
-        )
-        for row in rows:
-            if row.get("attachments"):
-                try:
-                    row["attachments"] = json.loads(row["attachments"])
-                except (TypeError, ValueError):
-                    row["attachments"] = []
-        return rows
-
-    def search_messages(self, query: str, limit: int = 50) -> list[dict[str, Any]]:
-        like = f"%{query}%"
-        return self._rows(
-            "SELECT m.*, c.title AS conversation_title FROM messages m "
-            "JOIN conversations c ON c.id = m.conversation_id "
-            "WHERE m.content LIKE ? ORDER BY m.created_at DESC LIMIT ?",
-            (like, limit),
-        )
-
-    # -- profile ---------------------------------------------------------------------
-
-    def get_profile(self, key: str, default: Any = None) -> Any:
-        row = self._row("SELECT value FROM profile WHERE key = ?", (key,))
-        if row is None:
-            return default
-        try:
-            return json.loads(row["value"])
-        except (TypeError, ValueError):
-            return row["value"]
-
-    def set_profile(self, key: str, value: Any) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO profile (key, value, updated_at) VALUES (?, ?, ?) "
-                "ON CONFLICT(key) DO UPDATE SET value = excluded.value, "
-                "updated_at = excluded.updated_at",
-                (key, json.dumps(value), utc_now()),
-            )
-
-    def all_profile(self) -> dict[str, Any]:
-        out: dict[str, Any] = {}
-        for row in self._rows("SELECT key, value FROM profile"):
-            try:
-                out[row["key"]] = json.loads(row["value"])
-            except (TypeError, ValueError):
-                out[row["key"]] = row["value"]
-        return out
-
-    def delete_profile(self, key: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM profile WHERE key = ?", (key,))
-
-    # -- people -----------------------------------------------------------------------
-
-    def add_person(
-        self,
-        name: str,
-        *,
-        relationship: str | None = None,
-        birthday: str | None = None,
-        allergies: str | None = None,
-        preferences: str | None = None,
-        gift_ideas: str | None = None,
-        notes: str | None = None,
-        person_id: str | None = None,
-    ) -> dict[str, Any]:
-        pid = person_id or new_id()
-        now = utc_now()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO people (id, name, relationship, birthday, allergies, "
-                "preferences, gift_ideas, notes, created_at, updated_at) "
-                "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                (
-                    pid,
-                    name,
-                    relationship,
-                    birthday,
-                    allergies,
-                    preferences,
-                    gift_ideas,
-                    notes,
-                    now,
-                    now,
-                ),
-            )
-        return self.get_person(pid)  # type: ignore[return-value]
-
-    def get_person(self, person_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM people WHERE id = ?", (person_id,))
-
-    def find_person(self, name: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM people WHERE lower(name) = lower(?) LIMIT 1", (name,))
-
-    def list_people(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM people ORDER BY name COLLATE NOCASE")
-
-    def update_person(self, person_id: str, **fields: Any) -> None:
-        allowed = {
-            "name",
-            "relationship",
-            "birthday",
-            "allergies",
-            "preferences",
-            "gift_ideas",
-            "notes",
-        }
-        updates = {k: v for k, v in fields.items() if k in allowed}
-        if not updates:
-            return
-        sets = ", ".join(f"{k} = ?" for k in updates)
-        params = (*updates.values(), utc_now(), person_id)
-        with self._write() as conn:
-            conn.execute(f"UPDATE people SET {sets}, updated_at = ? WHERE id = ?", params)
-
-    def delete_person(self, person_id: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM people WHERE id = ?", (person_id,))
-            conn.execute("DELETE FROM important_dates WHERE person_id = ?", (person_id,))
-
-    # -- important dates -----------------------------------------------------------------
-
-    def add_date(
-        self,
-        date: str,
-        label: str,
-        *,
-        recurring: bool = False,
-        reminder_days_before: int = 7,
-        person_id: str | None = None,
-        date_id: str | None = None,
-    ) -> dict[str, Any]:
-        did = date_id or new_id()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO important_dates (id, date, label, recurring, "
-                "reminder_days_before, person_id, created_at) VALUES (?, ?, ?, ?, ?, ?, ?)",
-                (
-                    did,
-                    date,
-                    label,
-                    1 if recurring else 0,
-                    reminder_days_before,
-                    person_id,
-                    utc_now(),
-                ),
-            )
-        return self._row("SELECT * FROM important_dates WHERE id = ?", (did,))  # type: ignore[return-value]
-
-    def list_dates(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM important_dates ORDER BY date")
-
-    def get_date(self, date_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM important_dates WHERE id = ?", (date_id,))
-
-    def update_date(self, date_id: str, **fields: Any) -> None:
-        allowed = {"date", "label", "recurring", "reminder_days_before", "person_id"}
-        updates = {key: value for key, value in fields.items() if key in allowed}
-        if "recurring" in updates:
-            updates["recurring"] = 1 if updates["recurring"] else 0
-        if not updates:
-            return
-        sets = ", ".join(f"{key} = ?" for key in updates)
-        with self._write() as conn:
-            conn.execute(
-                f"UPDATE important_dates SET {sets} WHERE id = ?",
-                (*updates.values(), date_id),
-            )
-
-    def delete_date(self, date_id: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM important_dates WHERE id = ?", (date_id,))
-
-    # -- memory journal ----------------------------------------------------------
-
-    def log_memory_journal(
-        self,
-        kind: str,
-        subject: str,
-        action: str,
-        value: Any = None,
-    ) -> None:
-        """Append one memory mutation to the journal (add|update|delete)."""
-        snapshot = json.dumps(value, ensure_ascii=False) if value is not None else None
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO memory_journal (kind, subject, action, value, created_at) "
-                "VALUES (?, ?, ?, ?, ?)",
-                (kind, subject, action, snapshot, utc_now()),
-            )
-
-    def list_memory_journal(self, limit: int = 100) -> list[dict[str, Any]]:
-        """Most recent journal entries first, newest on top."""
-        rows = self._rows("SELECT * FROM memory_journal ORDER BY id DESC LIMIT ?", (limit,))
-        for row in rows:
-            if row.get("value"):
-                with suppress(TypeError, ValueError):
-                    row["value"] = json.loads(row["value"])
-        return rows
-
-    # -- reminders ---------------------------------------------------------------------------
-
-    def add_reminder(
-        self,
-        text: str,
-        due_at: str,
-        *,
-        recurrence: str | None = None,
-        reminder_id: str | None = None,
-    ) -> dict[str, Any]:
-        rid = reminder_id or new_id()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO reminders (id, text, due_at, recurrence, created_at) "
-                "VALUES (?, ?, ?, ?, ?)",
-                (rid, text, due_at, recurrence, utc_now()),
-            )
-        return self._row("SELECT * FROM reminders WHERE id = ?", (rid,))  # type: ignore[return-value]
-
-    def list_reminders(self, include_completed: bool = False) -> list[dict[str, Any]]:
-        if include_completed:
-            return self._rows("SELECT * FROM reminders ORDER BY due_at")
-        return self._rows("SELECT * FROM reminders WHERE completed = 0 ORDER BY due_at")
-
-    def due_reminders(self, now: str) -> list[dict[str, Any]]:
-        return self._rows(
-            "SELECT * FROM reminders WHERE completed = 0 AND due_at <= ? "
-            "AND (snoozed_until IS NULL OR snoozed_until <= ?) ORDER BY due_at",
-            (now, now),
-        )
-
-    @staticmethod
-    def _resolve_reminder_id(conn: sqlite3.Connection, reminder_id: str) -> str | None:
-        """Resolve an exact ID or one unambiguous displayed 8-character prefix."""
-        exact = conn.execute("SELECT id FROM reminders WHERE id = ?", (reminder_id,)).fetchone()
-        if exact is not None:
-            return str(exact["id"])
-        if len(reminder_id) != 8 or any(
-            character not in "0123456789abcdefABCDEF" for character in reminder_id
-        ):
-            return None
-        matches = conn.execute(
-            "SELECT id FROM reminders WHERE substr(id, 1, 8) = ? LIMIT 2",
-            (reminder_id.lower(),),
-        ).fetchall()
-        return str(matches[0]["id"]) if len(matches) == 1 else None
-
-    def complete_reminder(self, reminder_id: str) -> bool:
-        with self._write() as conn:
-            resolved_id = self._resolve_reminder_id(conn, reminder_id)
-            if resolved_id is None:
-                return False
-            cursor = conn.execute(
-                "UPDATE reminders SET completed = 1 WHERE id = ?",
-                (resolved_id,),
-            )
-            return cursor.rowcount > 0
-
-    def snooze_reminder(self, reminder_id: str, until: str) -> bool:
-        with self._write() as conn:
-            resolved_id = self._resolve_reminder_id(conn, reminder_id)
-            if resolved_id is None:
-                return False
-            cursor = conn.execute(
-                "UPDATE reminders SET snoozed_until = ? WHERE id = ?",
-                (until, resolved_id),
-            )
-            return cursor.rowcount > 0
-
-    def delete_reminder(self, reminder_id: str) -> bool:
-        with self._write() as conn:
-            resolved_id = self._resolve_reminder_id(conn, reminder_id)
-            if resolved_id is None:
-                return False
-            cursor = conn.execute("DELETE FROM reminders WHERE id = ?", (resolved_id,))
-            return cursor.rowcount > 0
-
-    # -- shopping list -----------------------------------------------------------------------
-
-    def add_shopping_item(
-        self,
-        item: str,
-        *,
-        category: str = "Other",
-        quantity: str | None = None,
-        list_name: str = "Groceries",
-    ) -> dict[str, Any]:
-        iid = new_id()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO shopping_items (id, list_name, item, category, quantity, "
-                "created_at) VALUES (?, ?, ?, ?, ?, ?)",
-                (iid, list_name, item, category or "Other", quantity, utc_now()),
-            )
-        return self._row("SELECT * FROM shopping_items WHERE id = ?", (iid,))  # type: ignore[return-value]
-
-    def list_shopping_items(
-        self, list_name: str = "Groceries", include_checked: bool = True
-    ) -> list[dict[str, Any]]:
-        sql = "SELECT * FROM shopping_items WHERE list_name = ?"
-        if not include_checked:
-            sql += " AND checked = 0"
-        return self._rows(sql + " ORDER BY category, created_at", (list_name,))
-
-    def find_shopping_item(self, item: str, list_name: str = "Groceries") -> dict[str, Any] | None:
-        return self._row(
-            "SELECT * FROM shopping_items WHERE list_name = ? AND lower(item) = lower(?) LIMIT 1",
-            (list_name, item),
-        )
-
-    def check_shopping_item_by_name(self, item: str, list_name: str, checked: bool = True) -> int:
-        """Check/uncheck every row with the given name (duplicates included)."""
-        with self._write() as conn:
-            cur = conn.execute(
-                "UPDATE shopping_items SET checked = ? "
-                "WHERE list_name = ? AND lower(item) = lower(?)",
-                (1 if checked else 0, list_name, item),
-            )
-            return cur.rowcount
-
-    def delete_shopping_item_by_name(self, item: str, list_name: str) -> int:
-        """Delete every row with the given name (duplicates included)."""
-        with self._write() as conn:
-            cur = conn.execute(
-                "DELETE FROM shopping_items WHERE list_name = ? AND lower(item) = lower(?)",
-                (list_name, item),
-            )
-            return cur.rowcount
-
-    def clear_checked_shopping_items(self, list_name: str = "Groceries") -> int:
-        with self._write() as conn:
-            cur = conn.execute(
-                "DELETE FROM shopping_items WHERE list_name = ? AND checked = 1",
-                (list_name,),
-            )
-            return cur.rowcount
-
-    # -- budget ------------------------------------------------------------------------------
-
-    def add_expense(
-        self,
-        amount: float,
-        *,
-        category: str = "Other",
-        description: str | None = None,
-        spent_at: str | None = None,
-    ) -> dict[str, Any]:
-        eid = new_id()
-        day = spent_at or self._local_today()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO expenses (id, amount, category, description, spent_at, "
-                "created_at) VALUES (?, ?, ?, ?, ?, ?)",
-                (
-                    eid,
-                    amount,
-                    category or "Other",
-                    description,
-                    day,
-                    utc_now(),
-                ),
-            )
-        return self._row("SELECT * FROM expenses WHERE id = ?", (eid,))  # type: ignore[return-value]
-
-    def expenses_by_category(self, month: str) -> list[dict[str, Any]]:
-        return self._rows(
-            "SELECT category, SUM(amount) AS spent FROM expenses "
-            "WHERE spent_at LIKE ? GROUP BY category ORDER BY spent DESC",
-            (f"{month}%",),
-        )
-
-    def set_budget(self, category: str, monthly_limit: float) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO budgets (category, monthly_limit, updated_at) "
-                "VALUES (?, ?, ?) ON CONFLICT(category) DO UPDATE SET "
-                "monthly_limit = excluded.monthly_limit, updated_at = excluded.updated_at",
-                (category, monthly_limit, utc_now()),
-            )
-
-    def list_budgets(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM budgets ORDER BY category")
-
-    # -- health -------------------------------------------------------------------------------
-
-    def log_health(
-        self,
-        metric: str,
-        value: float,
-        *,
-        logged_on: str | None = None,
-        note: str | None = None,
-    ) -> dict[str, Any]:
-        # User-visible "today"/"this month" windows use the local calendar,
-        # so the default date must be local (not UTC) to stay consistent.
-        day = logged_on or self._local_today()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO health_logs (id, metric, value, logged_on, note, created_at) "
-                "VALUES (?, ?, ?, ?, ?, ?) ON CONFLICT(metric, logged_on) DO UPDATE SET "
-                "value = excluded.value, note = excluded.note",
-                (new_id(), metric, value, day, note, utc_now()),
-            )
-        return self._row(  # type: ignore[return-value]
-            "SELECT * FROM health_logs WHERE metric = ? AND logged_on = ?",
-            (metric, day),
-        )
-
-    def health_logs_since(self, since_day: str) -> list[dict[str, Any]]:
-        return self._rows(
-            "SELECT * FROM health_logs WHERE logged_on >= ? ORDER BY logged_on",
-            (since_day,),
-        )
-
-    def health_latest(self, metric: str) -> dict[str, Any] | None:
-        return self._row(
-            "SELECT * FROM health_logs WHERE metric = ? ORDER BY logged_on DESC LIMIT 1",
-            (metric,),
-        )
-
-    # -- automations ------------------------------------------------------------------------
-
-    def add_automation(
-        self,
-        name: str,
-        *,
-        description: str = "",
-        schedule: str = "",
-        action_type: str = "briefing",
-        action_config: Any = None,
-        enabled: bool = True,
-        delivery_channels: Any = None,
-        automation_id: str | None = None,
-        timezone_name: str = "UTC",
-        schedule_json: dict[str, Any] | None = None,
-        next_run_at: str | None = None,
-        plan_id: str | None = None,
-        plan_version: int | None = None,
-    ) -> dict[str, Any]:
-        aid = automation_id or new_id()
-        now = utc_now()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO automations (id, name, description, schedule, action_type, "
-                "action_config, enabled, delivery_channels, created_at, timezone, "
-                "schedule_json, next_run_at, plan_id, plan_version, routine_status, updated_at) "
-                "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                (
-                    aid,
-                    name,
-                    description,
-                    schedule,
-                    action_type,
-                    json.dumps(action_config) if action_config is not None else None,
-                    1 if enabled else 0,
-                    json.dumps(delivery_channels) if delivery_channels is not None else None,
-                    now,
-                    timezone_name,
-                    json.dumps(schedule_json) if schedule_json is not None else None,
-                    next_run_at,
-                    plan_id,
-                    plan_version,
-                    "enabled" if enabled else "paused",
-                    now,
-                ),
-            )
-        return self._row("SELECT * FROM automations WHERE id = ?", (aid,))  # type: ignore[return-value]
-
-    def list_automations(self, enabled_only: bool = False) -> list[dict[str, Any]]:
-        sql = "SELECT * FROM automations"
-        if enabled_only:
-            sql += " WHERE enabled = 1"
-        return self._rows(sql + " ORDER BY created_at")
-
-    def toggle_automation(self, automation_id: str, enabled: bool) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "UPDATE automations SET enabled = ?, routine_status = ?, updated_at = ?, "
-                # Re-enabling schedules from now: a stale next_run_at must not
-                # fire (or skip) instantly on resume.
-                "next_run_at = CASE WHEN ? = 1 THEN NULL ELSE next_run_at END "
-                "WHERE id = ?",
-                (
-                    1 if enabled else 0,
-                    "enabled" if enabled else "paused",
-                    utc_now(),
-                    1 if enabled else 0,
-                    automation_id,
-                ),
-            )
-
-    def mark_routine_result(
-        self, automation_id: str, *, success: bool, error: str | None = None
-    ) -> None:
-        now = utc_now()
-        with self._write() as conn:
-            self._mark_routine_result_with(conn, automation_id, success=success, now=now)
-
-    @staticmethod
-    def _mark_routine_result_with(conn, automation_id: str, *, success: bool, now: str) -> None:
-        if success:
-            conn.execute(
-                "UPDATE automations SET last_run = ?, last_success_at = ?, "
-                "consecutive_failures = 0, updated_at = ? WHERE id = ?",
-                (now, now, now, automation_id),
-            )
-        else:
-            conn.execute(
-                "UPDATE automations SET last_failure_at = ?, "
-                "consecutive_failures = consecutive_failures + 1, updated_at = ? WHERE id = ?",
-                (now, now, automation_id),
-            )
-
-    def get_automation(self, automation_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM automations WHERE id = ?", (automation_id,))
-
-    def update_automation(self, automation_id: str, **fields: Any) -> dict[str, Any]:
-        allowed = {
-            "name",
-            "description",
-            "schedule",
-            "timezone",
-            "schedule_json",
-            "next_run_at",
-            "missed_run_policy",
-            "plan_id",
-            "plan_version",
-            "routine_status",
-            "enabled",
-            "action_config",
-        }
-        updates = {key: value for key, value in fields.items() if key in allowed}
-        if not updates:
-            row = self.get_automation(automation_id)
-            if row is None:
-                raise ValueError("routine not found")
-            return row
-        for key in ("schedule_json", "action_config"):
-            if key in updates and not isinstance(updates[key], str):
-                updates[key] = json.dumps(updates[key])
-        updates["updated_at"] = utc_now()
-        assignments = ", ".join(f"{key} = ?" for key in updates)
-        with self._write() as conn:
-            cursor = conn.execute(
-                f"UPDATE automations SET {assignments} WHERE id = ?",
-                (*updates.values(), automation_id),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("routine not found")
-        return self.get_automation(automation_id)  # type: ignore[return-value]
-
-    def set_routine_shared_delivery(
-        self, automation_id: str, delivery: dict[str, Any] | None
-    ) -> dict[str, Any]:
-        with self._write() as conn:
-            cursor = conn.execute(
-                """UPDATE automations SET shared_delivery=?, shared_delivery_status=?,
-                shared_delivery_event_id=NULL, shared_delivery_error=NULL, updated_at=? WHERE id=?""",
-                (
-                    json.dumps(delivery, sort_keys=True) if delivery else None,
-                    "ready" if delivery else None,
-                    utc_now(),
-                    automation_id,
-                ),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("routine not found")
-        return self.get_automation(automation_id)  # type: ignore[return-value]
-
-    def record_routine_shared_delivery(
-        self, automation_id: str, *, status: str, event_id: str | None, error: str | None
-    ) -> None:
-        if status not in {"ready", "pending", "acknowledged", "rejected"}:
-            raise ValueError("invalid shared delivery status")
-        with self._write() as conn:
-            conn.execute(
-                """UPDATE automations SET shared_delivery_status=?,
-                shared_delivery_event_id=?, shared_delivery_error=?, updated_at=? WHERE id=?""",
-                (status, event_id, error, utc_now(), automation_id),
-            )
-
-    def delete_automation(self, automation_id: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM automations WHERE id = ?", (automation_id,))
-
-    # -- task checklists ---------------------------------------------------------------
-
-    def require_conversation_review(
-        self, conversation_id: str, reasons: list[str]
-    ) -> dict[str, Any]:
-        """Persist a review-first gate until an approved plan is claimed."""
-        normalized = list(
-            dict.fromkeys(
-                str(reason or "").strip() for reason in reasons if str(reason or "").strip()
-            )
-        )
-        if not normalized:
-            raise ValueError("A conversation review gate needs at least one reason.")
-        declared_at = utc_now()
-        with self._write_immediate() as conn:
-            if (
-                conn.execute(
-                    "SELECT 1 FROM conversations WHERE id = ?", (conversation_id,)
-                ).fetchone()
-                is None
-            ):
-                raise ValueError("This conversation no longer exists.")
-            conn.execute(
-                "INSERT INTO conversation_review_gates "
-                "(conversation_id, reasons_json, declared_at) VALUES (?, ?, ?) "
-                "ON CONFLICT(conversation_id) DO UPDATE SET "
-                "reasons_json = excluded.reasons_json, declared_at = excluded.declared_at",
-                (conversation_id, json.dumps(normalized), declared_at),
-            )
-        return {
-            "conversation_id": conversation_id,
-            "reasons": normalized,
-            "declared_at": declared_at,
-        }
-
-    def get_conversation_review_gate(self, conversation_id: str) -> dict[str, Any] | None:
-        row = self._row(
-            "SELECT conversation_id, reasons_json, declared_at "
-            "FROM conversation_review_gates WHERE conversation_id = ?",
-            (conversation_id,),
-        )
-        if row is None:
-            return None
-        try:
-            reasons = json.loads(str(row["reasons_json"]))
-        except (TypeError, json.JSONDecodeError):
-            reasons = []
-        return {
-            "conversation_id": str(row["conversation_id"]),
-            "reasons": reasons if isinstance(reasons, list) else [],
-            "declared_at": str(row["declared_at"]),
-        }
+    # -- shared status vocabulary (checklists and run tasks) ------------------
 
     _CHECKLIST_STEP_STATUSES = frozenset(
         {"pending", "in_progress", "completed", "blocked", "skipped", "failed"}
     )
     _CHECKLIST_TERMINAL_STATUSES = frozenset({"completed", "blocked", "failed", "cancelled"})
-
-    @staticmethod
-    def _checklist_task_with(conn: sqlite3.Connection, checklist: sqlite3.Row) -> dict[str, Any]:
-        steps = [
-            dict(row)
-            for row in conn.execute(
-                "SELECT * FROM task_checklist_steps WHERE checklist_id = ? ORDER BY ordinal",
-                (checklist["id"],),
-            ).fetchall()
-        ]
-        return {
-            "id": str(checklist["id"]),
-            "conversation_id": str(checklist["conversation_id"]),
-            "source": "checklist",
-            "status": str(checklist["status"]),
-            "revision": int(checklist["revision"]),
-            "title": str(checklist["goal"]),
-            "completed_count": sum(1 for step in steps if step["status"] == "completed"),
-            "total_count": len(steps),
-            "current_step_key": checklist["current_step_key"],
-            "steps": [
-                {
-                    "key": str(step["step_key"]),
-                    "title": str(step["title"]),
-                    "status": str(step["status"]),
-                    "summary": step["summary"],
-                    "error_message": step["error_message"],
-                }
-                for step in steps
-            ],
-            "created_at": checklist["created_at"],
-            "completed_at": checklist["completed_at"],
-        }
-
-    def create_task_checklist(
-        self,
-        *,
-        conversation_id: str,
-        goal: str,
-        steps: list[dict[str, str]],
-        review_plan_id: str | None = None,
-        review_plan_version: int | None = None,
-    ) -> dict[str, Any]:
-        goal = str(goal or "").strip()
-        if not goal:
-            raise ValueError("A task checklist needs a goal.")
-        if not isinstance(steps, list) or not 3 <= len(steps) <= 7:
-            raise ValueError("A task checklist needs 3 to 7 steps.")
-
-        normalized: list[tuple[str, str]] = []
-        keys: set[str] = set()
-        for ordinal, step in enumerate(steps):
-            if not isinstance(step, dict):
-                raise ValueError(f"Checklist step {ordinal + 1} must be an object.")
-            key = str(step.get("key") or "").strip()
-            title = str(step.get("title") or "").strip()
-            if not key or not title:
-                raise ValueError(f"Checklist step {ordinal + 1} needs a stable key and title.")
-            if key in keys:
-                raise ValueError(f"Checklist step key '{key}' is duplicated.")
-            keys.add(key)
-            normalized.append((key, title))
-
-        checklist_id = new_id()
-        now = utc_now()
-        try:
-            with self._write_immediate() as conn:
-                if (
-                    conn.execute(
-                        "SELECT 1 FROM conversations WHERE id = ?", (conversation_id,)
-                    ).fetchone()
-                    is None
-                ):
-                    raise ValueError("This conversation no longer exists.")
-                conn.execute(
-                    "INSERT INTO task_checklists (id, conversation_id, goal, status, "
-                    "current_step_key, review_plan_id, review_plan_version, revision, "
-                    "created_at, updated_at) VALUES (?, ?, ?, 'active', NULL, ?, ?, 1, ?, ?)",
-                    (
-                        checklist_id,
-                        conversation_id,
-                        goal,
-                        review_plan_id,
-                        review_plan_version,
-                        now,
-                        now,
-                    ),
-                )
-                for ordinal, (key, title) in enumerate(normalized):
-                    conn.execute(
-                        "INSERT INTO task_checklist_steps (id, checklist_id, step_key, "
-                        "ordinal, title, status) VALUES (?, ?, ?, ?, ?, 'pending')",
-                        (new_id(), checklist_id, key, ordinal, title),
-                    )
-                row = conn.execute(
-                    "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-                ).fetchone()
-                assert row is not None
-                result = self._checklist_task_with(conn, row)
-        except sqlite3.IntegrityError as exc:
-            if "task_checklists.conversation_id" in str(exc):
-                raise ValueError("This conversation already has an active checklist.") from exc
-            raise
-        return result
-
-    def get_task_checklist(self, checklist_id: str) -> dict[str, Any] | None:
-        with self._lock:
-            row = self._conn.execute(
-                "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-            ).fetchone()
-            return self._checklist_task_with(self._conn, row) if row is not None else None
-
-    @staticmethod
-    def _require_checklist_revision(
-        checklist: sqlite3.Row | None, expected_revision: int
-    ) -> sqlite3.Row:
-        if checklist is None:
-            raise ValueError("Task checklist not found.")
-        if int(checklist["revision"]) != int(expected_revision):
-            raise ValueError("This task checklist changed; use its latest revision.")
-        if str(checklist["status"]) != "active":
-            raise ValueError("This task checklist is already finished.")
-        return checklist
-
-    def update_task_checklist(
-        self,
-        checklist_id: str,
-        *,
-        expected_revision: int,
-        step_key: str,
-        status: str,
-        summary: str | None = None,
-        error_message: str | None = None,
-        title: str | None = None,
-    ) -> dict[str, Any]:
-        status = str(status or "").strip()
-        if status not in self._CHECKLIST_STEP_STATUSES:
-            raise ValueError(f"Unknown checklist step status: {status}")
-        summary = str(summary).strip() if summary is not None else None
-        error_message = str(error_message).strip() if error_message is not None else None
-        title = str(title).strip() if title is not None else None
-        if status in {"blocked", "failed"} and not error_message:
-            raise ValueError(f"A {status} checklist step needs an error message.")
-        if title is not None and not title:
-            raise ValueError("A checklist step title must not be blank.")
-
-        now = utc_now()
-        with self._write_immediate() as conn:
-            self._require_checklist_revision(
-                conn.execute(
-                    "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-                ).fetchone(),
-                expected_revision,
-            )
-            step = conn.execute(
-                "SELECT * FROM task_checklist_steps WHERE checklist_id = ? AND step_key = ?",
-                (checklist_id, step_key),
-            ).fetchone()
-            if step is None:
-                raise ValueError("Checklist step not found.")
-            old_status = str(step["status"])
-            if old_status == "completed" and status != "completed":
-                raise ValueError("A completed checklist step cannot move backwards.")
-            if old_status in {"blocked", "failed", "skipped"} and status != old_status:
-                raise ValueError("A finished checklist step cannot move backwards.")
-            if title is not None and (old_status != "pending" or status != "pending"):
-                raise ValueError("Only a pending checklist step can be renamed.")
-
-            other_current = conn.execute(
-                "SELECT step_key FROM task_checklist_steps WHERE checklist_id = ? "
-                "AND status = 'in_progress' AND step_key != ? LIMIT 1",
-                (checklist_id, step_key),
-            ).fetchone()
-            if status == "in_progress" and other_current is not None:
-                raise ValueError("Only one checklist step can be in progress.")
-            if status in {"blocked", "failed"} and other_current is not None:
-                raise ValueError("Finish the current checklist step before ending another one.")
-
-            started_at = step["started_at"]
-            if status == "in_progress" and not started_at:
-                started_at = now
-            finished_at = now if status in {"completed", "blocked", "skipped", "failed"} else None
-            conn.execute(
-                "UPDATE task_checklist_steps SET title = COALESCE(?, title), status = ?, "
-                "summary = COALESCE(?, summary), "
-                "error_message = COALESCE(?, error_message), "
-                "started_at = ?, finished_at = ? "
-                "WHERE checklist_id = ? AND step_key = ?",
-                (
-                    title,
-                    status,
-                    summary,
-                    error_message,
-                    started_at,
-                    finished_at,
-                    checklist_id,
-                    step_key,
-                ),
-            )
-            parent_status = status if status in {"blocked", "failed"} else "active"
-            current_row = conn.execute(
-                "SELECT step_key FROM task_checklist_steps WHERE checklist_id = ? "
-                "AND status = 'in_progress' ORDER BY ordinal LIMIT 1",
-                (checklist_id,),
-            ).fetchone()
-            current_step_key = str(current_row["step_key"]) if current_row is not None else None
-            completed_at = now if parent_status != "active" else None
-            cursor = conn.execute(
-                "UPDATE task_checklists SET status = ?, current_step_key = ?, "
-                "revision = revision + 1, updated_at = ?, completed_at = ? "
-                "WHERE id = ? AND revision = ?",
-                (
-                    parent_status,
-                    current_step_key,
-                    now,
-                    completed_at,
-                    checklist_id,
-                    expected_revision,
-                ),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("This task checklist changed; use its latest revision.")
-            refreshed = conn.execute(
-                "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-            ).fetchone()
-            assert refreshed is not None
-            return self._checklist_task_with(conn, refreshed)
-
-    def complete_task_checklist(
-        self, checklist_id: str, *, expected_revision: int
-    ) -> dict[str, Any]:
-        now = utc_now()
-        with self._write_immediate() as conn:
-            self._require_checklist_revision(
-                conn.execute(
-                    "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-                ).fetchone(),
-                expected_revision,
-            )
-            unfinished = conn.execute(
-                "SELECT 1 FROM task_checklist_steps WHERE checklist_id = ? "
-                "AND status NOT IN ('completed', 'skipped') LIMIT 1",
-                (checklist_id,),
-            ).fetchone()
-            if unfinished is not None:
-                raise ValueError("Complete or skip every checklist step before finishing the task.")
-            cursor = conn.execute(
-                "UPDATE task_checklists SET status = 'completed', current_step_key = NULL, "
-                "revision = revision + 1, updated_at = ?, completed_at = ? "
-                "WHERE id = ? AND revision = ?",
-                (now, now, checklist_id, expected_revision),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("This task checklist changed; use its latest revision.")
-            row = conn.execute(
-                "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-            ).fetchone()
-            assert row is not None
-            return self._checklist_task_with(conn, row)
-
-    def cancel_task_checklist(
-        self,
-        checklist_id: str,
-        *,
-        expected_revision: int,
-        reason: str | None = None,
-    ) -> dict[str, Any]:
-        now = utc_now()
-        reason = str(reason).strip() if reason is not None else None
-        with self._write_immediate() as conn:
-            self._require_checklist_revision(
-                conn.execute(
-                    "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-                ).fetchone(),
-                expected_revision,
-            )
-            conn.execute(
-                "UPDATE task_checklist_steps SET status = 'skipped', error_message = ?, "
-                "finished_at = ? WHERE checklist_id = ? AND status = 'in_progress'",
-                (reason, now, checklist_id),
-            )
-            cursor = conn.execute(
-                "UPDATE task_checklists SET status = 'cancelled', current_step_key = NULL, "
-                "revision = revision + 1, updated_at = ?, completed_at = ? "
-                "WHERE id = ? AND revision = ?",
-                (now, now, checklist_id, expected_revision),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("This task checklist changed; use its latest revision.")
-            row = conn.execute(
-                "SELECT * FROM task_checklists WHERE id = ?", (checklist_id,)
-            ).fetchone()
-            assert row is not None
-            return self._checklist_task_with(conn, row)
 
     @staticmethod
     def _normalize_run_step_status(status: str) -> str:
@@ -2829,778 +1709,6 @@ class CollieDB:
     def list_run_steps(self, run_id: str) -> list[dict[str, Any]]:
         return self._rows("SELECT * FROM run_steps WHERE run_id = ? ORDER BY ordinal", (run_id,))
 
-    # -- approval rules and requests ---------------------------------------------------
-
-    def add_approval_rule(
-        self,
-        *,
-        action: str,
-        resource_pattern: str,
-        effect: str,
-        scope_type: str,
-        scope_value: str | None = None,
-        created_by: str = "user",
-        expires_at: str | None = None,
-    ) -> dict[str, Any]:
-        rule_id = new_id()
-        now = utc_now()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO approval_rules (id, action, resource_pattern, effect, "
-                "scope_type, scope_value, created_by, expires_at, created_at, updated_at) "
-                "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                (
-                    rule_id,
-                    action,
-                    resource_pattern,
-                    effect,
-                    scope_type,
-                    scope_value,
-                    created_by,
-                    expires_at,
-                    now,
-                    now,
-                ),
-            )
-        return self._row("SELECT * FROM approval_rules WHERE id = ?", (rule_id,))  # type: ignore[return-value]
-
-    def list_approval_rules(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM approval_rules ORDER BY created_at DESC")
-
-    def delete_approval_rule(self, rule_id: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM approval_rules WHERE id = ?", (rule_id,))
-
-    def create_approval_request(
-        self,
-        *,
-        action: str,
-        resource: str,
-        risk: str,
-        display: dict[str, Any],
-        run_id: str | None = None,
-        conversation_id: str | None = None,
-        step_id: str | None = None,
-        tool_call_id: str | None = None,
-    ) -> dict[str, Any]:
-        request_id = new_id()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO approval_requests (id, run_id, conversation_id, step_id, "
-                "tool_call_id, action, resource, risk, display_json, status, requested_at) "
-                "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, 'pending', ?)",
-                (
-                    request_id,
-                    run_id,
-                    conversation_id,
-                    step_id,
-                    tool_call_id,
-                    action,
-                    resource,
-                    risk,
-                    self._canonical_json(display),
-                    utc_now(),
-                ),
-            )
-        return self._row("SELECT * FROM approval_requests WHERE id = ?", (request_id,))  # type: ignore[return-value]
-
-    def list_pending_approvals(self) -> list[dict[str, Any]]:
-        rows = self._rows(
-            "SELECT * FROM approval_requests WHERE status = 'pending' ORDER BY requested_at"
-        )
-        for row in rows:
-            row["display"] = json.loads(row["display_json"])
-        return rows
-
-    def resolve_approval_request(
-        self, request_id: str, resolution: str, rule_id: str | None = None
-    ) -> dict[str, Any]:
-        with self._write() as conn:
-            cursor = conn.execute(
-                "UPDATE approval_requests SET status = 'resolved', resolved_at = ?, "
-                "resolution = ?, rule_id = ? WHERE id = ? AND status = 'pending'",
-                (utc_now(), resolution, rule_id, request_id),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("approval request is no longer pending")
-        return self._row("SELECT * FROM approval_requests WHERE id = ?", (request_id,))  # type: ignore[return-value]
-
-    # -- services ----------------------------------------------------------------------------
-
-    def upsert_service(
-        self,
-        service_id: str,
-        *,
-        name: str,
-        provider: str,
-        auth_type: str = "oauth",
-        status: str = "disconnected",
-        account_info: str | None = None,
-        last_error: str | None = None,
-    ) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO services (id, name, provider, auth_type, status, "
-                "account_info, connected_at, last_error) VALUES (?, ?, ?, ?, ?, ?, ?, ?) "
-                "ON CONFLICT(id) DO UPDATE SET status = excluded.status, "
-                "account_info = excluded.account_info, last_error = excluded.last_error, "
-                "connected_at = CASE WHEN excluded.status = 'connected' "
-                "THEN excluded.connected_at ELSE services.connected_at END",
-                (
-                    service_id,
-                    name,
-                    provider,
-                    auth_type,
-                    status,
-                    account_info,
-                    utc_now() if status == "connected" else None,
-                    last_error,
-                ),
-            )
-
-    def get_service(self, service_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM services WHERE id = ?", (service_id,))
-
-    def list_services(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM services ORDER BY name COLLATE NOCASE")
-
-    # -- connectors --------------------------------------------------------------------------
-
-    def upsert_connector_connection(
-        self,
-        connection_id: str,
-        *,
-        provider_id: str,
-        driver: str,
-        auth_type: str,
-        status: str,
-        display_name: str | None = None,
-        account_label: str | None = None,
-        granted_scopes: list[str] | None = None,
-        enabled_capabilities: list[str] | None = None,
-        enabled_tools: list[str] | None = None,
-        tool_policy: dict[str, Any] | None = None,
-        remote_account_id: str | None = None,
-        last_verified_at: str | None = None,
-        last_error_code: str | None = None,
-        last_error_message: str | None = None,
-        definition_id: str | None = None,
-    ) -> dict[str, Any]:
-        now = utc_now()
-        with self._write() as conn:
-            existing = conn.execute(
-                "SELECT definition_id, provider_id, driver FROM connector_connections WHERE id = ?",
-                (connection_id,),
-            ).fetchone()
-            if (
-                existing
-                and definition_id is not None
-                and existing["definition_id"] is not None
-                and definition_id != existing["definition_id"]
-            ):
-                raise ValueError("An installed connector definition cannot be replaced in place.")
-            definition_id = (
-                definition_id
-                or (
-                    str(existing["definition_id"])
-                    if existing and existing["definition_id"]
-                    else None
-                )
-                or f"def_{connection_id}"
-            )
-            associated = conn.execute(
-                "SELECT provider_id, driver FROM connector_definitions WHERE id = ?",
-                (definition_id,),
-            ).fetchone()
-            if associated and (
-                associated["provider_id"] not in (None, provider_id)
-                or associated["driver"] != driver
-            ):
-                raise ValueError("The connector definition does not match this account.")
-            conn.execute(
-                "INSERT OR IGNORE INTO connector_definitions "
-                "(id, provider_id, recipe_id, driver, transport, auth_strategy, provenance, "
-                "config_json, unresolved, created_at, updated_at) "
-                "VALUES (?, ?, ?, ?, ?, ?, 'curated', '{}', 1, ?, ?)",
-                (
-                    definition_id,
-                    provider_id,
-                    provider_id,
-                    driver,
-                    "api" if driver.endswith("api") else "streamable_http",
-                    "oauth"
-                    if auth_type == "oauth"
-                    else ("none" if auth_type == "none" else "token"),
-                    now,
-                    now,
-                ),
-            )
-            conn.execute(
-                """
-                INSERT INTO connector_connections (
-                    id, provider_id, display_name, account_label, driver, auth_type,
-                    status, granted_scopes_json, enabled_capabilities_json,
-                    enabled_tools_json, tool_policy_json, remote_account_id,
-                    connected_at, updated_at, last_verified_at, last_error_code,
-                    last_error_message, definition_id, credential_ref
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                ON CONFLICT(id) DO UPDATE SET
-                    display_name = COALESCE(excluded.display_name, display_name),
-                    account_label = COALESCE(excluded.account_label, account_label),
-                    status = excluded.status,
-                    granted_scopes_json = COALESCE(
-                        excluded.granted_scopes_json, granted_scopes_json
-                    ),
-                    enabled_capabilities_json = COALESCE(
-                        excluded.enabled_capabilities_json, enabled_capabilities_json
-                    ),
-                    enabled_tools_json = COALESCE(
-                        excluded.enabled_tools_json, enabled_tools_json
-                    ),
-                    tool_policy_json = COALESCE(
-                        excluded.tool_policy_json, tool_policy_json
-                    ),
-                    remote_account_id = COALESCE(
-                        excluded.remote_account_id, remote_account_id
-                    ),
-                    connected_at = CASE
-                        WHEN excluded.status = 'connected'
-                        THEN COALESCE(connected_at, excluded.connected_at)
-                        ELSE connected_at
-                    END,
-                    updated_at = excluded.updated_at,
-                    last_verified_at = COALESCE(
-                        excluded.last_verified_at, last_verified_at
-                    ),
-                    last_error_code = excluded.last_error_code,
-                    last_error_message = excluded.last_error_message
-                """,
-                (
-                    connection_id,
-                    provider_id,
-                    display_name,
-                    account_label,
-                    driver,
-                    auth_type,
-                    status,
-                    json.dumps(granted_scopes) if granted_scopes is not None else None,
-                    (
-                        json.dumps(enabled_capabilities)
-                        if enabled_capabilities is not None
-                        else None
-                    ),
-                    json.dumps(enabled_tools) if enabled_tools is not None else None,
-                    json.dumps(tool_policy) if tool_policy is not None else None,
-                    remote_account_id,
-                    now if status == "connected" else None,
-                    now,
-                    last_verified_at,
-                    last_error_code,
-                    last_error_message,
-                    definition_id,
-                    f"connector:{connection_id}",
-                ),
-            )
-            if enabled_tools is not None:
-                conn.execute(
-                    "UPDATE connector_tool_cache SET enabled = CASE WHEN EXISTS "
-                    "(SELECT 1 FROM json_each(?) WHERE value = '*' "
-                    "OR value = remote_tool_name) THEN 1 ELSE 0 END WHERE connection_id = ?",
-                    (json.dumps(enabled_tools), connection_id),
-                )
-        return self.get_connector_connection(connection_id)  # type: ignore[return-value]
-
-    def get_connector_connection(self, connection_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM connector_connections WHERE id = ?", (connection_id,))
-
-    def list_connector_connections(self, provider_id: str | None = None) -> list[dict[str, Any]]:
-        if provider_id:
-            return self._rows(
-                "SELECT * FROM connector_connections WHERE provider_id = ? "
-                "ORDER BY updated_at DESC",
-                (provider_id,),
-            )
-        return self._rows("SELECT * FROM connector_connections ORDER BY updated_at DESC")
-
-    def delete_connector_connection(self, connection_id: str) -> None:
-        with self._write() as conn:
-            row = conn.execute(
-                "SELECT definition_id FROM connector_connections WHERE id = ?", (connection_id,)
-            ).fetchone()
-            conn.execute("DELETE FROM connector_connections WHERE id = ?", (connection_id,))
-            if row and row["definition_id"]:
-                conn.execute(
-                    "DELETE FROM connector_definitions WHERE id = ? AND NOT EXISTS "
-                    "(SELECT 1 FROM connector_connections WHERE definition_id = ?)",
-                    (row["definition_id"], row["definition_id"]),
-                )
-
-    def save_connector_definition(self, definition: Any) -> dict[str, Any]:
-        """Persist a validated, secret-free installed definition snapshot."""
-        from collie_core.connectors.models import InstalledConnectorDefinition
-
-        if not isinstance(definition, InstalledConnectorDefinition):
-            definition = InstalledConnectorDefinition.from_dict(definition)
-        value = definition.to_dict()
-        config = {
-            key: value[key]
-            for key in (
-                "endpoint",
-                "oauth_registration",
-                "client_id",
-                "scopes",
-                "trusted_hosts",
-                "tool_overrides",
-            )
-            if value[key] not in (None, [], "")
-        }
-        now = utc_now()
-        with self._write() as conn:
-            current = conn.execute(
-                "SELECT * FROM connector_definitions WHERE id = ?", (definition.id,)
-            ).fetchone()
-            if current is not None and not current["unresolved"]:
-                current_config = json.loads(str(current["config_json"]))
-                immutable = (
-                    current["provider_id"],
-                    current["recipe_id"],
-                    current["recipe_version"],
-                    current["driver"],
-                    current["transport"],
-                    current["auth_strategy"],
-                    current["provenance"],
-                    current_config,
-                )
-                proposed = (
-                    definition.provider_id,
-                    definition.recipe_id,
-                    definition.recipe_version,
-                    definition.driver.value,
-                    definition.transport.value,
-                    definition.auth_strategy.value,
-                    definition.provenance.value,
-                    config,
-                )
-                if immutable != proposed:
-                    raise ValueError("Installed connector definitions are immutable.")
-                return dict(current)
-            conn.execute(
-                "INSERT INTO connector_definitions (id, provider_id, recipe_id, recipe_version, "
-                "driver, transport, auth_strategy, provenance, config_json, unresolved, "
-                "created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?) "
-                "ON CONFLICT(id) DO UPDATE SET provider_id=excluded.provider_id, "
-                "recipe_id=excluded.recipe_id, recipe_version=excluded.recipe_version, "
-                "driver=excluded.driver, transport=excluded.transport, "
-                "auth_strategy=excluded.auth_strategy, provenance=excluded.provenance, "
-                "config_json=excluded.config_json, unresolved=excluded.unresolved, "
-                "updated_at=excluded.updated_at",
-                (
-                    definition.id,
-                    definition.provider_id,
-                    definition.recipe_id,
-                    definition.recipe_version,
-                    definition.driver.value,
-                    definition.transport.value,
-                    definition.auth_strategy.value,
-                    definition.provenance.value,
-                    json.dumps(config, sort_keys=True),
-                    int(definition.unresolved),
-                    now,
-                    now,
-                ),
-            )
-        return self.get_connector_definition(definition.id)  # type: ignore[return-value]
-
-    def get_connector_definition(self, definition_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM connector_definitions WHERE id = ?", (definition_id,))
-
-    def list_connector_definitions(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM connector_definitions ORDER BY created_at, id")
-
-    def advance_connector_operation(self, connection_id: str, expected_revision: int) -> int:
-        """Claim the next lifecycle operation; reject absent or stale accounts."""
-        with self._write() as conn:
-            cursor = conn.execute(
-                "UPDATE connector_connections SET operation_revision = operation_revision + 1, "
-                "updated_at = ? WHERE id = ? AND operation_revision = ?",
-                (utc_now(), connection_id, expected_revision),
-            )
-            if cursor.rowcount != 1:
-                raise ValueError("The connection changed or no longer exists.")
-        return expected_revision + 1
-
-    def replace_connector_tools(self, connection_id: str, tools: list[dict[str, Any]]) -> None:
-        with self._write() as conn:
-            account = conn.execute(
-                "SELECT enabled_tools_json FROM connector_connections WHERE id = ?",
-                (connection_id,),
-            ).fetchone()
-            configured = (
-                json.loads(account["enabled_tools_json"])
-                if (account and account["enabled_tools_json"] is not None)
-                else None
-            )
-            conn.execute(
-                "DELETE FROM connector_tool_cache WHERE connection_id = ?",
-                (connection_id,),
-            )
-            conn.executemany(
-                "INSERT INTO connector_tool_cache "
-                "(connection_id, remote_tool_name, schema_hash, annotations_json, "
-                "risk, discovered_at, tool_identity, description, input_schema_json, enabled) "
-                "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-                [
-                    (
-                        connection_id,
-                        str(tool["name"]),
-                        str(tool["schema_hash"]),
-                        json.dumps(tool.get("annotations") or {}),
-                        str(tool["risk"]),
-                        utc_now(),
-                        f"{len(connection_id)}:{connection_id}:"
-                        f"{len(str(tool['name']))}:{tool['name']}",
-                        tool.get("description"),
-                        json.dumps(tool.get("input_schema"))
-                        if tool.get("input_schema") is not None
-                        else None,
-                        int(
-                            bool(
-                                tool.get(
-                                    "enabled",
-                                    (
-                                        configured is None
-                                        or "*" in configured
-                                        or str(tool["name"]) in configured
-                                    ),
-                                )
-                            )
-                        ),
-                    )
-                    for tool in tools
-                ],
-            )
-
-    def list_connector_tools(self, connection_id: str) -> list[dict[str, Any]]:
-        return self._rows(
-            "SELECT * FROM connector_tool_cache WHERE connection_id = ? ORDER BY remote_tool_name",
-            (connection_id,),
-        )
-
-    # -- subagents ----------------------------------------------------------------------------
-
-    def upsert_subagent(
-        self,
-        name: str,
-        *,
-        description: str = "",
-        system_prompt: str = "",
-        filename: str = "",
-        execution_posture: str = "read_only",
-        subagent_id: str | None = None,
-    ) -> dict[str, Any]:
-        sid = subagent_id or new_id()
-        now = utc_now()
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO subagents (id, name, description, system_prompt, filename, "
-                "execution_posture, created_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?) "
-                "ON CONFLICT(id) DO UPDATE SET name = excluded.name, "
-                "description = excluded.description, system_prompt = excluded.system_prompt, "
-                "filename = excluded.filename, execution_posture = excluded.execution_posture, "
-                "updated_at = excluded.updated_at",
-                (
-                    sid,
-                    name,
-                    description,
-                    system_prompt,
-                    filename,
-                    execution_posture,
-                    now,
-                    now,
-                ),
-            )
-        return self._row("SELECT * FROM subagents WHERE id = ?", (sid,))  # type: ignore[return-value]
-
-    def list_subagents(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM subagents ORDER BY name COLLATE NOCASE")
-
-    def delete_subagent(self, subagent_id: str) -> None:
-        with self._write() as conn:
-            conn.execute("DELETE FROM subagents WHERE id = ?", (subagent_id,))
-
-    # -- providers & usage -----------------------------------------------------------------------
-
-    _PROVIDER_SETTING_KEYS = (
-        "provider.auth",
-        "provider.name",
-        "provider.model",
-        "provider.api_base",
-        "provider.secret_name",
-    )
-
-    def snapshot_provider_configuration(self, provider_id: str) -> dict[str, Any]:
-        """Capture exactly the provider state needed for a failed candidate rollback."""
-        with self._lock:
-            provider_row = self._conn.execute(
-                "SELECT * FROM providers WHERE id = ?", (provider_id,)
-            ).fetchone()
-            default_row = self._conn.execute(
-                "SELECT id FROM providers WHERE is_default = 1 LIMIT 1"
-            ).fetchone()
-            raw_settings: dict[str, str | None] = {}
-            for key in self._PROVIDER_SETTING_KEYS:
-                row = self._conn.execute(
-                    "SELECT value FROM settings WHERE key = ?", (key,)
-                ).fetchone()
-                raw_settings[key] = row["value"] if row is not None else None
-            return {
-                "provider_id": provider_id,
-                "provider": dict(provider_row) if provider_row is not None else None,
-                "default_provider_id": default_row["id"] if default_row is not None else None,
-                "settings": raw_settings,
-            }
-
-    def configure_provider_candidate_record(
-        self,
-        provider_id: str,
-        *,
-        name: str,
-        auth_type: str,
-        model: str | None,
-        runtime_name: str,
-        protocol: str,
-        api_base: str | None,
-        secret_name: str,
-    ) -> dict[str, Any]:
-        """Tentatively select a provider row and its settings in one transaction."""
-        now = utc_now()
-        settings = {
-            "provider.auth": auth_type,
-            "provider.name": runtime_name,
-            "provider.model": model,
-            "provider.api_base": api_base,
-            "provider.secret_name": secret_name,
-        }
-        with self._write() as conn:
-            conn.execute("UPDATE providers SET is_default = 0")
-            conn.execute(
-                "INSERT INTO providers (id, name, auth_type, is_default, model, created_at, "
-                "runtime_name, protocol, api_base, secret_name) "
-                "VALUES (?, ?, ?, 1, ?, ?, ?, ?, ?, ?) "
-                "ON CONFLICT(id) DO UPDATE SET name = excluded.name, "
-                "auth_type = excluded.auth_type, is_default = 1, model = excluded.model, "
-                "runtime_name = excluded.runtime_name, protocol = excluded.protocol, "
-                "api_base = excluded.api_base, secret_name = excluded.secret_name",
-                (
-                    provider_id,
-                    name,
-                    auth_type,
-                    model,
-                    now,
-                    runtime_name,
-                    protocol,
-                    api_base,
-                    secret_name,
-                ),
-            )
-            for key, value in settings.items():
-                conn.execute(
-                    "INSERT INTO settings (key, value) VALUES (?, ?) "
-                    "ON CONFLICT(key) DO UPDATE SET value = excluded.value",
-                    (key, json.dumps(value)),
-                )
-        return self.get_provider(provider_id)  # type: ignore[return-value]
-
-    def restore_provider_configuration(self, snapshot: dict[str, Any]) -> None:
-        """Restore a snapshot made by :meth:`snapshot_provider_configuration`."""
-        provider_id = str(snapshot["provider_id"])
-        previous = snapshot.get("provider")
-        with self._write() as conn:
-            if previous is None:
-                conn.execute("DELETE FROM usage WHERE provider_id = ?", (provider_id,))
-                conn.execute("DELETE FROM providers WHERE id = ?", (provider_id,))
-            else:
-                columns = (
-                    "id",
-                    "name",
-                    "auth_type",
-                    "is_default",
-                    "model",
-                    "created_at",
-                    "last_used",
-                    "runtime_name",
-                    "protocol",
-                    "api_base",
-                    "secret_name",
-                )
-                values = tuple(previous.get(column) for column in columns)
-                conn.execute(
-                    "INSERT INTO providers (" + ", ".join(columns) + ") "
-                    "VALUES (" + ", ".join("?" for _ in columns) + ") "
-                    "ON CONFLICT(id) DO UPDATE SET "
-                    + ", ".join(
-                        f"{column} = excluded.{column}" for column in columns if column != "id"
-                    ),
-                    values,
-                )
-            conn.execute("UPDATE providers SET is_default = 0")
-            default_provider_id = snapshot.get("default_provider_id")
-            if default_provider_id:
-                conn.execute(
-                    "UPDATE providers SET is_default = 1 WHERE id = ?",
-                    (str(default_provider_id),),
-                )
-            for key, raw_value in dict(snapshot.get("settings") or {}).items():
-                if raw_value is None:
-                    conn.execute("DELETE FROM settings WHERE key = ?", (key,))
-                else:
-                    conn.execute(
-                        "INSERT INTO settings (key, value) VALUES (?, ?) "
-                        "ON CONFLICT(key) DO UPDATE SET value = excluded.value",
-                        (key, raw_value),
-                    )
-
-    def upsert_provider(
-        self,
-        provider_id: str,
-        *,
-        name: str,
-        auth_type: str,
-        model: str | None = None,
-        runtime_name: str | None = None,
-        protocol: str = "openai",
-        api_base: str | None = None,
-        secret_name: str | None = None,
-        is_default: bool = False,
-    ) -> None:
-        with self._write() as conn:
-            if is_default:
-                conn.execute("UPDATE providers SET is_default = 0")
-            conn.execute(
-                "INSERT INTO providers (id, name, auth_type, is_default, model, created_at, "
-                "runtime_name, protocol, api_base, secret_name) "
-                "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?) "
-                "ON CONFLICT(id) DO UPDATE SET name = excluded.name, "
-                "auth_type = excluded.auth_type, is_default = excluded.is_default, "
-                "model = excluded.model, runtime_name = excluded.runtime_name, "
-                "protocol = excluded.protocol, api_base = excluded.api_base, "
-                "secret_name = excluded.secret_name",
-                (
-                    provider_id,
-                    name,
-                    auth_type,
-                    1 if is_default else 0,
-                    model,
-                    utc_now(),
-                    runtime_name or name,
-                    protocol,
-                    api_base,
-                    secret_name or name,
-                ),
-            )
-
-    def get_provider(self, provider_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM providers WHERE id = ?", (provider_id,))
-
-    def list_providers(self) -> list[dict[str, Any]]:
-        return self._rows("SELECT * FROM providers ORDER BY created_at")
-
-    def default_provider(self) -> dict[str, Any] | None:
-        row = self._row("SELECT * FROM providers WHERE is_default = 1 LIMIT 1")
-        if row:
-            return row
-        return self._row("SELECT * FROM providers ORDER BY created_at LIMIT 1")
-
-    def set_default_provider(self, provider_id: str) -> None:
-        with self._write() as conn:
-            conn.execute("UPDATE providers SET is_default = 0")
-            conn.execute("UPDATE providers SET is_default = 1 WHERE id = ?", (provider_id,))
-
-    def delete_provider(self, provider_id: str) -> None:
-        with self._write() as conn:
-            # usage rows carry an FK to providers — remove them first or the
-            # delete itself fails while leaving both rows behind.
-            conn.execute("DELETE FROM usage WHERE provider_id = ?", (provider_id,))
-            conn.execute("DELETE FROM providers WHERE id = ?", (provider_id,))
-
-    def record_usage(self, provider_id: str, *, messages: int = 1, tokens: int = 0) -> None:
-        day = datetime.now(UTC).strftime("%Y-%m-%d")
-        with self._write() as conn:
-            conn.execute(
-                "INSERT INTO usage (id, provider_id, date, message_count, token_count) "
-                "VALUES (?, ?, ?, ?, ?) "
-                "ON CONFLICT(provider_id, date) DO UPDATE SET "
-                "message_count = usage.message_count + excluded.message_count, "
-                "token_count = usage.token_count + excluded.token_count",
-                (new_id(), provider_id, day, messages, tokens),
-            )
-
-    def usage_this_month(self, provider_id: str | None = None) -> dict[str, int]:
-        month_prefix = datetime.now(UTC).strftime("%Y-%m") + "%"
-        if provider_id:
-            rows = self._rows(
-                "SELECT COALESCE(SUM(message_count),0) AS messages, "
-                "COALESCE(SUM(token_count),0) AS tokens FROM usage "
-                "WHERE date LIKE ? AND provider_id = ?",
-                (month_prefix, provider_id),
-            )
-        else:
-            rows = self._rows(
-                "SELECT COALESCE(SUM(message_count),0) AS messages, "
-                "COALESCE(SUM(token_count),0) AS tokens FROM usage WHERE date LIKE ?",
-                (month_prefix,),
-            )
-        row = rows[0] if rows else {"messages": 0, "tokens": 0}
-        return {"messages": int(row["messages"]), "tokens": int(row["tokens"])}
-
-    def _increment_product_metrics(
-        self,
-        conn: sqlite3.Connection,
-        started_at: str | None,
-        *,
-        runs: int = 0,
-        interactive_runs: int = 0,
-        tool_calls: int = 0,
-    ) -> None:
-        if not self._product_metrics_enabled:
-            return
-        day = datetime.fromisoformat(started_at or utc_now()).astimezone(UTC).date().isoformat()
-        conn.execute(
-            "INSERT INTO product_metrics_daily (day, runs, interactive_runs, tool_calls) "
-            "VALUES (?, ?, ?, ?) ON CONFLICT(day) DO UPDATE SET "
-            "runs = runs + excluded.runs, "
-            "interactive_runs = interactive_runs + excluded.interactive_runs, "
-            "tool_calls = tool_calls + excluded.tool_calls",
-            (day, runs, interactive_runs, tool_calls),
-        )
-
-    def product_metrics(self) -> dict[str, Any]:
-        """Bounded, content-free cumulative counters; never export raw run records.
-
-        Source identity and counters deliberately do not travel in cloud backups.
-        Clearing local data rotates the source so new counts do not collide with
-        high-water marks already accepted by the server.
-        """
-        if not self._product_metrics_enabled:
-            return {}
-        today = datetime.now(UTC).date()
-        cutoff = (today - timedelta(days=34)).isoformat()
-        with self._write_immediate() as conn:
-            row = conn.execute("SELECT id FROM product_metrics_source").fetchone()
-            source = row["id"] if row else str(uuid.uuid4())
-            if not row:
-                conn.execute("INSERT INTO product_metrics_source VALUES (?)", (source,))
-            conn.execute("DELETE FROM product_metrics_daily WHERE day < ?", (cutoff,))
-            days = [
-                dict(row)
-                for row in conn.execute(
-                    "SELECT day, runs, interactive_runs, tool_calls FROM product_metrics_daily "
-                    "WHERE day >= ? AND day <= ? ORDER BY day",
-                    (cutoff, today.isoformat()),
-                )
-            ]
-        return {"source_id": source, "days": days}
-
     # -- run records (telemetry) ----------------------------------------------------------------
 
     def record_turn_event(
@@ -3859,95 +1967,6 @@ class CollieDB:
             {"tool_name": row["tool_name"], "status": row["status"], "count": int(row["count"])}
             for row in rows
         ]
-
-    # -- artifact versions (Gardener rollback rail) ---------------------------------------------
-
-    def snapshot_artifact(
-        self,
-        artifact_type: str,
-        artifact_key: str,
-        before_text: str,
-        after_text: str,
-        diff_text: str,
-        evidence: Any = None,
-        source: str = "user",
-    ) -> dict[str, Any]:
-        """Append a version row for an artifact edit; returns the new row."""
-        with self._write_immediate() as conn:
-            row = conn.execute(
-                "SELECT MAX(version) AS v FROM artifact_versions "
-                "WHERE artifact_type = ? AND artifact_key = ?",
-                (artifact_type, artifact_key),
-            ).fetchone()
-            version = int(row["v"] or 0) + 1
-            version_id = new_id()
-            conn.execute(
-                "INSERT INTO artifact_versions (id, artifact_type, artifact_key, "
-                "version, before_text, after_text, diff_text, evidence_json, "
-                "source, status, created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, "
-                "'applied', ?)",
-                (
-                    version_id,
-                    artifact_type,
-                    artifact_key,
-                    version,
-                    before_text,
-                    after_text,
-                    diff_text,
-                    json.dumps(evidence, ensure_ascii=False) if evidence is not None else None,
-                    source,
-                    utc_now(),
-                ),
-            )
-            return {
-                "id": version_id,
-                "artifact_type": artifact_type,
-                "artifact_key": artifact_key,
-                "version": version,
-                "status": "applied",
-            }
-
-    def latest_artifact_version(self, artifact_type: str, artifact_key: str) -> int:
-        with self._lock:
-            row = self._conn.execute(
-                "SELECT MAX(version) AS v FROM artifact_versions "
-                "WHERE artifact_type = ? AND artifact_key = ?",
-                (artifact_type, artifact_key),
-            ).fetchone()
-            return int(row["v"] or 0) if row else 0
-
-    def list_artifact_versions(
-        self,
-        artifact_type: str | None = None,
-        artifact_key: str | None = None,
-        limit: int = 100,
-    ) -> list[dict[str, Any]]:
-        """List artifact versions, most recent first, with optional filters."""
-        clauses: list[str] = []
-        params: list[Any] = []
-        if artifact_type:
-            clauses.append("artifact_type = ?")
-            params.append(artifact_type)
-        if artifact_key:
-            clauses.append("artifact_key = ?")
-            params.append(artifact_key)
-        where = f"WHERE {' AND '.join(clauses)}" if clauses else ""
-        params.append(limit)
-        return self._rows(
-            f"SELECT * FROM artifact_versions {where} "
-            "ORDER BY created_at DESC, version DESC LIMIT ?",
-            tuple(params),
-        )
-
-    def get_artifact_version(self, version_id: str) -> dict[str, Any] | None:
-        return self._row("SELECT * FROM artifact_versions WHERE id = ?", (version_id,))
-
-    def mark_artifact_rolled_back(self, version_id: str) -> None:
-        with self._write() as conn:
-            conn.execute(
-                "UPDATE artifact_versions SET status = 'rolled_back' WHERE id = ?",
-                (version_id,),
-            )
 
     # -- data management ----------------------------------------------------------------------------
